+++ client/js/helpers/get-info.js
@@ -1,6 +1,6 @@
 export const getInfo = async (path, parse = true) => {
-    const {useProxy, proxy, hosts, useHost} = globalThis.config
-    const SERVER_ADDRESS = useProxy ? `${proxy}?server=${useHost}&request=${path}` : `http://${hosts[useHost]}/${path}`
+    const {useProxy, proxy, hosts, useHost, useHttps = false} = globalThis.config
+    const SERVER_ADDRESS = useProxy ? `${proxy}?server=${useHost}&request=${path}` : `${useHttps ? "https" : "http"}://${hosts[useHost]}/${path}`
 
     try {
         const result = await fetch(`${SERVER_ADDRESS}`)